Update 'dafni get workflow'
