+++ dafni_cli/workflow/workflow.py
@@ -1,239 +1,403 @@
-import datetime as dt
+from dataclasses import dataclass
+from datetime import date, datetime
+from typing import ClassVar, List, Optional
 
 import click
 from dateutil import parser
 
+from dafni_cli.api.parser import ParserBaseObject, ParserParam, parse_datetime
 from dafni_cli.api.session import DAFNISession
-from dafni_cli.api.workflows_api import get_workflow  # get_workflow_metadata_dict
+from dafni_cli.api.workflows_api import get_workflow
 from dafni_cli.auth import Auth
 from dafni_cli.consts import CONSOLE_WIDTH, TAB_SPACE
 from dafni_cli.utils import print_json, prose_print
-from dafni_cli.workflow.workflow_metadata import WorkflowMetadata
 
 
-class Workflow:
-    """
-    Contains information about a DAFNI workflow.
-    The information (as attributes) for a workflow can be populated from a dictionary, or a DAFNI workflow UUID.
-
-    Methods:
-        get_attributes_from_dict(dict): populates attributes from the workflow dictionary from the DAFNI API
-        get_attributes_from_id(session (DAFNISession), id (str)): populates attributes from the workflow version ID by calling DAFNI API.
-        get_metadata(session (DAFNISession)): After details have been obtained, populate metadata attributes.
-        filter_by_date(key (str), date (str)): calculates whether the workflow was created/published before a date.
-        output_details(): Prints key information of workflow to console.
-        output_metadata(): Prints key information of workflow metadata to console.
+@dataclass
+class WorkflowMetadata(ParserBaseObject):
+    """Dataclass representing a DAFNI workflows's metadata
 
     Attributes:
-        api_version: The version of the API that created the data
-        auth: credentials used to retrieve the data
-        creation_date: Date and time the workflow was created
-        description: More detailed information on the workflow
-        display_name: Name of the workflow used in the web app
-        id: ID used to identify the specific workflow version
-        instances: A list of the workflow version run instances
-        kind: The type of asset (will be 'W', workflow)
-        metadata: a sub-group containing the workflow metadata:
-            description: More detailed information on the workflow
-            display_name: Name of the workflow used in the web app
-            name: DAFNI workflow name
-            publisher: Organisation publishing the model
-            summary: One-line summary of what the workflow does
-            status: ?
-        name: DAFNI workflow name
-        owner: UUID of the workflow owner
-        parameter_sets:
-        parent: UUID of the workflow's parent
-        publication_date: Date and time the workflow was published
-        publisher: Organisation publishing the workflow
-        spec: JSON description of the workflow steps
-        summary: One-line summary of what the workflow does
-        version_history: A list of all versions of the workflow
-        version_message: Message attached when the workflow was updated to this workflow version
-        version_tags: Any tags created by the publisher for this version
-
-        dictionary: Full description of the workflow as retrieved from the API
-        metadata_obj: WorkflowMetadata object containing metadata for the workflow
+        description (str): A rich description of the Model's function
+        display_name (str): The display name of the Model
+        name (str): Name of the model
+        publisher_id (str): The id of the person or organisation who has
+                            published the Model
+        summary (str): A short summary of the Model's function
     """
 
-    def __init__(self, identifier=None):
-        # Attributes that are also workflow dictionary keys
-        self.api_version = None
-        self.auth = Auth()
-        self.creation_date = None
-        self.display_name = None
-        self.description = None  # present??
-        self.id = identifier
-        self.instances = None
-        self.kind = None
-        self.metadata = None
-        self.name = None
-        self.owner = None
-        self.parameter_sets = None
-        self.parent = None
-        self.publication_date = None
-        self.publisher = None
-        self.spec = None
-        self.summary = None
-        self.version_history = None
-        self.version_message = None
-        self.version_tags = None
+    description: str
+    display_name: str
+    name: str
+    publisher_id: str
+    summary: str
 
-        # Set of workflow dictionary keys, for use in validation functions
-        workflow_attributes = vars(self).keys()
-        self.workflow_attributes = set(workflow_attributes)
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("description", "description", str),
+        ParserParam("display_name", "display_name", str),
+        ParserParam("name", "name", str),
+        ParserParam("publisher_id", "publisher", str),
+        ParserParam("summary", "summary", str),
+    ]
 
-        # Attributes that are not also workflow dictionary keys
-        self.dictionary = None
-        self.metadata_obj = None
 
-        return
+# TODO: Unify with ModelVersion
+@dataclass
+class WorkflowVersion(ParserBaseObject):
+    """Dataclass containing information on a historic version of a workflow
 
-    def set_attributes_from_dict(self, workflow_dict: dict):
-        """
-        Attempts to store workflow attributes from a dictionary returned from the DAFNI API.
-        Not all of the attributes need be present in workflow_dict.
+    Attributes:
+        version_id (str): ID of the version
+        version_message (str): Message labelling the model version
+        version_tags (List[str]): Version tags e.g. 'latest'
+        publication_date: Date and time this version was published
+    """
+
+    version_id: str
+    version_message: str
+    version_tags: List[str]
+    publication_date: datetime
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("version_id", "id", str),
+        ParserParam("version_message", "version_message", str),
+        ParserParam("version_tags", "version_tags"),
+        ParserParam("publication_date", "publication_date", parse_datetime),
+    ]
+
+
+# TODO: Unify with ModelAuth
+@dataclass
+class WorkflowAuth(ParserBaseObject):
+    """Dataclass representing the access the user has for a DAFNI workflow
+
+    Attributes:
+        view (bool): View access
+        read (bool): Read access
+        update (bool): Update access
+        destroy (bool): Deletion access
+        reason (str): Reason user has access to view this model
+
+
+        The following are only present for the /workflow/<version_id> endpoint
+        (but are guaranteed for it)
+        --------
+        asset_id (Optional[str]): ID of the workflow
+
+
+        The following are only present for the /workflows endpoint
+        (but are guaranteed for it)
+        --------
+        role_id (Optional[str]): Role ID of the user
+        name (Optional[str]): Name associated with the auth type
+    """
+
+    view: bool
+    read: bool
+    update: bool
+    destroy: bool
+    reason: str
+
+    asset_id: Optional[str] = None
+    name: Optional[str] = None
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("view", "view"),
+        ParserParam("read", "read"),
+        ParserParam("update", "update"),
+        ParserParam("destroy", "destroy"),
+        ParserParam("reason", "reason", str),
+        ParserParam("asset_id", "asset_id", str),
+    ]
+
+
+@dataclass
+class WorkflowInstanceParameterSet(ParserBaseObject):
+    """Dataclass representing the information gathered about the parameter set
+       a workflow instance was executed with
+
+    Attributes:
+        parameter_set_id (str): ID of the parameter set
+        display_name (str): Display name of the parameter set
+    """
+
+    parameter_set_id: str
+    display_name: str
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("parameter_set_id", "id", str),
+        ParserParam("display_name", "display_name", str),
+    ]
+
+
+@dataclass
+class WorkflowInstanceWorkflowVersion(ParserBaseObject):
+    """Dataclass representing the information gathered about the workflow
+       version a workflow instance was executed with
+
+    Attributes:
+        version_id (str): Version ID of the workflow
+        version_message (str): Message labelling the workflow version
+    """
+
+    version_id: str
+    version_message: str
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("version_id", "id", str),
+        ParserParam("version_message", "version_message", str),
+    ]
+
+
+# TODO: Check parsing mid workflow execution (expect finished time to be None?)
+@dataclass
+class WorkflowInstance(ParserBaseObject):
+    """Dataclass representing a workflow instance (an execution of a DAFNI
+       workflow)
+
+    Attributes:
+        instance_id (str): ID of this instance
+        submission_time (datetime): Date and time the instance was submitted
+                                    for execution
+        finished_time (datetime): Date and time the instance finished
+                                  execution
+        overall_status (str): Status of the overall workflow execution e.g.
+                              'Succeeded'
+        parameter_set (WorkflowInstanceParameterSet): Information on the
+                            parameter set the instance used
+        workflow_version (WorkflowInstanceWorkflowVersion): Information on the
+                            wowkflow version the instance used
+    """
+
+    instance_id: str
+    submission_time: datetime
+    finished_time: datetime
+    overall_status: str
+    parameter_set: WorkflowInstanceParameterSet
+    workflow_version: WorkflowInstanceWorkflowVersion
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("instance_id", "instance_id", str),
+        ParserParam("submission_time", "submission_time", str),
+        ParserParam("finished_time", "finished_time", str),
+        ParserParam("overall_status", "overall_status", str),
+        ParserParam("parameter_set", "parameter_set", WorkflowInstanceParameterSet),
+        ParserParam(
+            "workflow_version", "workflow_version", WorkflowInstanceWorkflowVersion
+        ),
+    ]
+
+
+@dataclass
+class WorkflowParameterSetMetadata(ParserBaseObject):
+    """Dataclass representing the metadata of a parameter set in a DAFNI
+       workflow
+
+    Attributes:
+        description (str): A rich description of the Model's function
+        display_name (str): The display name of the Model
+        name (str): Name of the model
+        publisher (str): The name of the person or organisation who has
+                         published the Model
+        workflow_version_id (str): Version id of the workflow
+    """
+
+    description: str
+    display_name: str
+    name: str
+    publisher: str
+    workflow_version_id: str
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("description", "description", str),
+        ParserParam("display_name", "display_name", str),
+        ParserParam("name", "name", str),
+        ParserParam("publisher", "publisher", str),
+        ParserParam("workflow_version_id", "workflow_version", str),
+    ]
+
+
+@dataclass
+class WorkflowParameterSet(ParserBaseObject):
+    """Dataclass representing a parameter set of a DAFNI workflow
+
+    Attributes:
+        parameter_set_id (str): ID of the parameter set
+        owner_id (str): ID of the parameter set owner
+        creation_date (datetime): Date and time the parameter set was created
+        publication_date (datetime): Date and time the parameter set was
+                                     published
+        kind (str): Type of DAFNI object (should be "P" for parameter set)
+        api_version (str): Version of the DAFNI API used to retrieve the
+                           parameter set data
+        spec (dict): Specification of the parameter set (contains information
+                     on the dataslots and parameters)
+        metadata (WorkflowParameterSetMetadata): Metadata of the parameter set
+    """
+
+    parameter_set_id: str
+    owner_id: str
+    creation_date: datetime
+    publication_date: datetime
+    kind: str
+    api_version: str
+    # TODO: Left as a dict for now, would just need its own parsing function
+    spec: dict
+    metadata: WorkflowParameterSetMetadata
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("parameter_set_id", "id", str),
+        ParserParam("owner_id", "owner", str),
+        ParserParam("creation_date", "creation_date", parse_datetime),
+        ParserParam("publication_date", "publication_date", parse_datetime),
+        ParserParam("kind", "kind", str),
+        ParserParam("api_version", "api_version", str),
+        ParserParam("spec", "spec"),
+        ParserParam("metadata", "metadata", WorkflowParameterSetMetadata),
+    ]
+
+
+@dataclass
+class Workflow(ParserBaseObject):
+    """Dataclass representing a DAFNI workflow
+
+    Attributes:
+        workflow_id (str): Workflow ID
+        metadata (WorkflowMetadata): Metadata of the workflow
+        version_history (List[WorkflowVersion]): Full version history of the
+                                                 workflow
+        auth (WorkflowAuth): Authentication credentials giving the permissions
+                             the current user has on the model
+        instances (List[WorkflowInstance]): Workflow instances (executions of
+                                            this workflow)
+        parameter_sets (List[WorkflowInstance]): Parameter sets that can be
+                                                 run with this workflow
+        api_version (Optional[str]): Version of the DAFNI API used to retrieve
+                                     the model data
+        kind (str): Type of DAFNI object (should be "W" for workflow)
+        creation_date (datetime): Date and time the workflow was created
+        publication_date (datetime): Date and time the workflow was published
+        owner_id (str): ID of the model owner
+        version_tags (List[str]): Any tags created by the publisher for this
+                                  workflow version
+        version_message (str): Message attached when the workflow was updated
+                               to this model version
+        spec (dict): Specification of the workflow (contains the steps of the
+                     workflow)
+        parent_id (str): Parent workflow ID
+    """
+
+    workflow_id: str
+    metadata: WorkflowMetadata
+    version_history: List[WorkflowVersion]
+    auth: WorkflowAuth
+    instances: List[WorkflowInstance]
+    parameter_sets: List[WorkflowParameterSet]
+    api_version: str
+    kind: str
+    creation_date: datetime
+    publication_date: datetime
+    owner_id: str
+    version_tags: List[str]
+    version_message: str
+    # TODO: Left as a dict for now, would just need its own parsing function
+    spec: dict
+    parent_id: str
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("workflow_id", "id", str),
+        ParserParam("metadata", "metadata", WorkflowMetadata),
+        ParserParam("version_history", "version_history", WorkflowVersion),
+        ParserParam("auth", "auth", WorkflowAuth),
+        ParserParam("instances", "instances", WorkflowInstance),
+        ParserParam("parameter_sets", "parameter_sets", WorkflowParameterSet),
+        ParserParam("api_version", "api_version", str),
+        ParserParam("kind", "kind", str),
+        ParserParam("creation_date", "creation_date", parse_datetime),
+        ParserParam("publication_date", "publication_date", parse_datetime),
+        ParserParam("owner_id", "owner", str),
+        ParserParam("version_tags", "version_tags"),
+        ParserParam("version_message", "version_message", str),
+        ParserParam("spec", "spec"),
+        ParserParam("parent_id", "parent", str),
+    ]
+
+    # TODO: Unify with Model
+    def filter_by_date(self, key: str, date_str: str) -> bool:
+        """Returns whether a particular date is greater than or equal to the
+           creation/publication date of this model.
 
         Args:
-            workflow_dict (dict): Dictionary returned from DAFNI API at /workflows endpoints
+            key (str): Key for which date to check must be either 'creation'
+                       or 'publication'
+            date_str (str): Date for which models are to be filtered on - format
+                            DD/MM/YYYY
 
         Returns:
-            a set of attributes of Workflow that could not be set as they were not present in workflow_dict
+            bool: Whether the given date is greater than or equal to the
+                  chosen date
         """
-        workflow_attributes = self.workflow_attributes
-        # special_attributes = {"auth", "creation_date", "publication_date"}
-        # workflow_attributes = self.workflow_attributes.difference_update(special_attributes)
-
-        missing_workflow_attributes = set()
-        for key in workflow_attributes:
-            try:
-                setattr(self, key, workflow_dict[key])
-            except:
-                missing_workflow_attributes.add(key)
-            pass
-
-        # Special treatment. If these keys are not in the missing attributes set, then
-        # they have already been added above and we just need to ignore the exception
-        try:
-            self.auth = Auth(workflow_dict["auth"])
-        except:
-            pass
-
-        try:
-            self.creation_date = parser.isoparse(workflow_dict["creation_date"])
-        except:
-            pass
-
-        try:
-            self.publication_date = parser.isoparse(workflow_dict["publication_date"])
-        except:
-            pass
-
-        #  If workflow_dict is not a dictionary then an exception will have already been raised
-        self.dictionary = workflow_dict
-
-        return missing_workflow_attributes
-
-    def get_attributes_from_id(self, session: str, version_id_string: str):
-        """
-        Retrieve workflow attributes from the DAFNI API using the /workflows/<version-id> endpoint.
-
-        Args:
-            session (DAFNISession): User session
-            version_id_string (str): Version ID of the workflow
-        """
-        workflow_dict = get_workflow(session, version_id_string)
-        self.set_attributes_from_dict(workflow_dict)
-        # TODO: Check: Version message key appears on single workflow API response, but not list of all workflows response
-        self.version_message = workflow_dict["version_message"]
-        return
-
-    def get_metadata(self, session: DAFNISession):
-        """
-        Retrieve metadata for the workflow using the workflow details.
-
-        Args:
-            session (DAFNISession): User session
-        """
-        #        metadata_dict = get_workflow_metadata_dict(session, self.id)
-        self.metadata_obj = WorkflowMetadata(self.dictionary)
-        return
-
-    def filter_by_date(self, key: str, date: str) -> bool:
-        """Filters workflows based on the date given as an option.
-        Args:
-            key (str): Key for WORKFLOW_DICT in which date is contained
-            date (str): Date for which workflows are to be filtered on: format DD/MM/YYYY
-
-        Returns:
-            bool: Whether to display the workflow based on the filter
-        """
-        day, month, year = date.split("/")
-        date = dt.date(int(year), int(month), int(day))
+        day, month, year = date_str.split("/")
+        date_val = date(int(year), int(month), int(day))
         if key.lower() == "creation":
-            return self.creation_date.date() >= date
-        elif key.lower() == "publication":
-            return self.publication_date.date() >= date
-        else:
-            raise KeyError("Key should be CREATION or PUBLICATION")
+            return self.creation_date.date() >= date_val
+        if key.lower() == "publication":
+            return self.publication_date.date() >= date_val
+        raise KeyError("Key should be 'creation' or 'publication'")
 
     def output_details(self, long: bool = False):
-        """
-        Prints relevant workflow attributes to command line
+        """Prints relevant workflow attributes to command line
+
+        Args:
+            long (bool): Whether to print with the (potentially long)
+                         description (ignored if description is None)
         """
         click.echo(
             "Name: "
-            + self.display_name
+            + self.metadata.display_name
             + TAB_SPACE
             + "ID: "
-            + self.id
+            + self.workflow_id
             + TAB_SPACE
             + "Date: "
             + self.creation_date.date().strftime("%B %d %Y")
         )
-        click.echo("Summary: " + self.summary)
-        if long:
+        click.echo("Summary: " + self.metadata.summary)
+        if long and self.metadata.description is not None:
             click.echo("Description: ")
-            prose_print(self.description, CONSOLE_WIDTH)
+            prose_print(self.metadata.description, CONSOLE_WIDTH)
         click.echo("")
-        return
 
-    def output_metadata(self, json_flag: bool = False):
-        """
-        Prints the metadata for the workflow to command line.
+    def output_info(self):
+        """Prints information about the workflow to command line"""
 
-        Args:
-            json_flag (bool): Whether to print raw json or pretty print information. Defaults to False.
-        """
-        if not json_flag:
-            click.echo("Name: " + self.metadata["display_name"])
-            click.echo("Date: " + self.creation_date.strftime("%B %d %Y"))
-            click.echo("Summary: ")
-            click.echo(self.metadata["summary"])
-            prose_print(self.metadata["description"], CONSOLE_WIDTH)
-            click.echo("")
-            if self.metadata_obj.inputs:
-                click.echo("Input Parameters: ")
-                click.echo(self.metadata_obj.format_parameters())
-                click.echo("Input Data Slots: ")
-                click.echo(self.metadata_obj.format_dataslots())
-            if self.metadata_obj.outputs:
-                click.echo("Outputs: ")
-                click.echo(self.metadata_obj.format_outputs())
-        else:
-            print_json(self.dictionary)
-        return
+        click.echo("Name: " + self.metadata.display_name)
+        click.echo("Date: " + self.creation_date.strftime("%B %d %Y"))
+        click.echo("Summary: ")
+        click.echo(self.metadata.summary)
+        prose_print(self.metadata.description, CONSOLE_WIDTH)
+        click.echo("")
+        
+        # TODO: Update this so can view inputs and outputs?
+        # if self.metadata_obj.inputs:
+        #     click.echo("Input Parameters: ")
+        #     click.echo(self.metadata_obj.format_parameters())
+        #     click.echo("Input Data Slots: ")
+        #     click.echo(self.metadata_obj.format_dataslots())
+        # if self.metadata_obj.outputs:
+        #     click.echo("Outputs: ")
+        #     click.echo(self.metadata_obj.format_outputs())
 
     def output_version_details(self) -> str:
-        """
-        Prints version ID, display name, publication time and version message on one line
+        """Prints workflow ID, display name, publication time and version
+           message on one line
         """
         return (
             "ID: "
-            + self.id
+            + self.workflow_id
             + TAB_SPACE
             + "Name: "
-            + self.metadata["display_name"]
+            + self.metadata.display_name
             + TAB_SPACE
             + "Publication date: "
             + self.publication_date.date().strftime("%B %d %Y")
@@ -241,3 +405,20 @@ class Workflow:
             + "Version message: "
             + self.version_message
         )
+    
+    def output_version_history(self):
+        """Prints the version history for the workflow to the command line"""
+        for version in self.version_history:
+            click.echo(
+                "Name: "
+                + self.metadata.display_name
+                + TAB_SPACE
+                + "ID: "
+                + version.version_id
+                + TAB_SPACE
+                + "Date: "
+                + version.publication_date.strftime("%B %d %Y")
+            )
+            click.echo(f"Version message: {version.version_message}")
+            click.echo(f"Version tags: {', '.join(version.version_tags)}")
+            click.echo("")
